Stop rewriting same PlayStore RPC data to cache
Previously the code would load from cache and then rewrite the same data
to cache. This is a problem for concurrency because reading/writing from
a file will break when another process is doing the same thing.

This adds a new class, DiscoveryAPI, to coordinate access to the cache
file and it reduces collisions by only writing to the cache when there
isn't a file.